refactor: 增大图标清晰度 & 优化代码
