+++ pwa-assets.config.ts
@@ -5,5 +5,5 @@ export default defineConfig({
         preset: "2023",
     },
     preset,
-    images: ["public/icon-512.png"],
+    images: ["public/icon.png"],
 });
