+++ src/components/file/FileItem.ts
@@ -11,6 +11,7 @@ export class FileItem {
     progress: number | undefined = 0;
     hash: string | undefined = undefined;
     estimetedTime: number = 0;
+    checkSum: string | undefined = undefined;
 
     constructor(addTime: number, name: string) {
         this.addTime = addTime;
